lwmecps-gym-api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
motor>=3.0.0
wandb>=0.12.0
python-dotenv>=0.19.0
prometheus-client>=0.12.0
lwmecps-gym-core>=0.1.0 